% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomart_human}
\alias{biomart_human}
\alias{biomart_mouse}
\title{Biomart annotation (mouse)}
\format{A data frame with 55029 obs. of  7 variables. The data frame contains the gene ID, the gene name, gene description, the chromosome scaffold name, strand, gene start and stop basepairs.}
\usage{
biomart_mouse
}
\description{
Gene annotations as downloaded brom biomart database on 2018-09-14, Genome version GRCm38.p6.
}
\references{
\url{https://www.ensembl.org/biomart/}
}
\keyword{annotation}
