% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{compareGSEA}
\alias{compareGSEA}
\title{GO & KEGG enrichment across comparisons}
\usage{
compareGSEA(comparisons, organism, GeneSets = c("GO", "KEGG"),
  ontology = "BP", pCorrection = "bonferroni", pvalueCutoff = 0.05,
  qvalueCutoff = 0.05, showMax = 20)
}
\description{
GO & KEGG enrichment across comparisons
}
