% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Corr_Dist.R
\name{corr_function}
\alias{corr_function}
\title{Correlation & Distance}
\usage{
corr_function(sampleCor = cd_input, gene_anno = gene_annotation,
  plot_anno = plot_annotation, title = title, gene_type = "all",
  cluster_rows = F, cluster_cols = F, mean = F)
}
\arguments{
\item{input}{A dataframe containing the read counts of the genes for each sample. By default, "norm_anno" normalized read counts is used.}

\item{anno}{A sample table containing info for the samples}

\item{condition}{A character from sample table df defining the name of column in the table to use. "condition" is set by default.}
}
\value{
Returns a dataframe
}
\description{
This function provides the sample-to-sample correlation .
}
