% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\name{DEAnalysis}
\alias{DEAnalysis}
\title{DESeq2 differential testing}
\usage{
DEAnalysis(condition, alpha = 0.05, lfcThreshold = 0, sigFC = 2,
  multiple_testing = "IHW", independentFiltering = TRUE,
  shrinkage = TRUE, shrinkType = "normal")
}
\description{
Wrapper Function to perform DESeq2 differential testing
}
