% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\docType{class}
\name{DESeq2_analysis_object-class}
\alias{DESeq2_analysis_object-class}
\title{Specify structure of DESeq2_analysis_object}
\description{
Specify structure of DESeq2_analysis_object
}
