% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Corr_Dist.R
\name{dist_function}
\alias{dist_function}
\title{This function provides the sample-to-sample distances .}
\usage{
dist_function(sampleDist = cd_input, gene_anno = gene_annotation,
  plot_anno = plot_annotation, title = title, gene_type = "all",
  mean = F)
}
\arguments{
\item{sampleDist}{input}

\item{gene_anno}{}

\item{sample_table}{A sample table containing info for the samples}
}
\value{
Returns a dataframe
}
\description{
This function provides the sample-to-sample distances .
}
