% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sampletable}
\alias{example_sampletable}
\title{Example sample table referring to the example dataset}
\format{An object of class \code{data.frame} with 12 rows and 5 columns.}
\usage{
example_sampletable
}
\description{
Example sample table referring to the example dataset
}
\references{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE92852}
}
\keyword{datasets}
