% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_hs}
\alias{GO_hs}
\alias{biomart_human}
\title{List of GO terms (human)}
\format{A data frame with 1310914 obs. of  5 variables.
 \describe{
  \item{GENEID}{Ontology term}
  \item{SYMBOL}{Entrez ID}
  \item{TERM}{GO term description}
  \item{DEFINITION}{GO term definition}
  \item{DOMAIN}{GO domain. biological_process, cellular_component or molecular_function}
  }}
\usage{
GO_hs

biomart_human
}
\description{
A dataset containing GO term description, the definition and the domain for human gene IDs.

A dataset containing GO term description, the definition and the domain for mouse gene IDs.
}
\references{
original file GO_hg38p12_ensembl181121.txt

original file GO_mm38p12_ensembl181121.txt

\url{https://www.ensembl.org/biomart/}
}
\keyword{annotation}
\keyword{gene}
\keyword{ontology}
