% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_hs_1}
\alias{GO_hs_1}
\title{List of GO terms (human) - Part 1}
\format{A data frame
 \describe{
  \item{GENEID}{Ontology term}
  \item{SYMBOL}{Entrez ID}
  \item{TERM}{GO term description}
  \item{DEFINITION}{GO term definition}
  \item{DOMAIN}{GO domain. biological_process, cellular_component or molecular_function}
  }}
\usage{
GO_hs_1
}
\description{
A dataset containing GO term description, the definition and the domain for human gene IDs.
Needs to be combined with GO_hs_2 to get the full dataframe.
}
\references{
original file GO_hg38p12_ensembl181121.txt
}
\keyword{gene}
\keyword{ontology}
