% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_mm_1}
\alias{GO_mm_1}
\alias{biomart_human}
\title{List of GO terms (mouse) - Part 1}
\format{A data frame
 \describe{
  \item{GENEID}{Ontology term}
  \item{SYMBOL}{Entrez ID}
  \item{TERM}{GO term description}
  \item{DEFINITION}{GO term definition}
  \item{DOMAIN}{GO domain. biological_process, cellular_component or molecular_function}
  }
List of GO terms (mouse) - Part 2

A dataset containing GO term description, the definition and the domain for mouse gene IDs.
Needs to be combined with GO_mm_1 to get the full dataframe.}
\usage{
biomart_human
}
\description{
A dataset containing GO term description, the definition and the domain for mouse gene IDs.
Needs to be combined with GO_mm_2 to get the full dataframe.
}
\references{
original file GO_mm38p12_ensembl181121.txt

original file GO_mm38p12_ensembl181121.txt

\url{https://www.ensembl.org/biomart/}
}
\keyword{annotation}
\keyword{gene}
\keyword{ontology}
