% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{GSEA function}
\usage{
GSEA(comparison, organism, GeneSets = c("GO", "KEGG", "Hallmark",
  "cannonicalPathways", "Motifs", "ImmunoSignatures"), GOntology = "BP",
  pCorrection = "bonferroni", pvalueCutoff = 0.05,
  qvalueCutoff = 0.05, showMax = 20, font.size = 8)
}
\description{
GSEA function
}
