% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hallmark_genes}
\alias{hallmark_genes}
\title{Hallmark genes}
\format{A data frame
 \describe{
  \item{ont}{Pathway}
  \item{gene}{Entrez ID}
  }}
\usage{
hallmark_genes
}
\description{
Hallmark genes
}
\references{
This datasets corresponds to the GMT file h.all.v6.2.entrez.gmt \url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\keyword{genes}
\keyword{pathway}
