% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immuno_genes}
\alias{immuno_genes}
\title{Immuno genes}
\format{A data frame
 \describe{
  \item{ont}{Pathway}
  \item{gene}{Entrez ID}
  }}
\usage{
immuno_genes
}
\description{
Immuno genes
}
\references{
This datasets corresponds to the GMT file c7.all.v6.2.entrez.gmt \url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\keyword{datasets}
