% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG_hs}
\alias{KEGG_hs}
\title{KEGG terms human}
\format{A data frame}
\usage{
KEGG_hs
}
\description{
KEGG terms human
}
\references{
original filename KEGG_hg38_clusterProfiler181121.txt
}
\keyword{datasets}
