% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG_mm}
\alias{KEGG_mm}
\title{KEGG terms mouse}
\format{A data frame}
\usage{
KEGG_mm
}
\description{
KEGG terms and corresponding genes.
}
\references{
Original filename KEGG_mm10_clusterProfiler181121.txt
}
\keyword{datasets}
