% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batcheffects.R
\name{limmaBatchEffectRemoval}
\alias{limmaBatchEffectRemoval}
\title{Removes batch effects}
\usage{
limmaBatchEffectRemoval(input = dds_vst, batchfactor,
  batchfactor_2 = NULL, modelfactor)
}
\description{
Function to remove potential batch effects using the removeBatchEffect function from limma
}
