% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Means.R
\name{mean_function}
\alias{mean_function}
\title{Mean of replicates}
\usage{
mean_function(input = norm_anno, anno = sample_table,
  condition = "condition")
}
\arguments{
\item{input}{A dataframe containing the read counts of the genes for each sample. By default, "norm_anno" normalized read counts is used.}

\item{anno}{A sample table containing info for the samples}

\item{condition}{A character from sample table df defining the name of column in the table to use. "condition" is set by default.}
}
\value{
Returns a dataframe
}
\description{
This function provides the calculation of the average of replicates per condition.
}
