% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Means.R
\name{mean_sample_definition}
\alias{mean_sample_definition}
\title{Generate mean sample table}
\usage{
mean_sample_definition(input = mean_sample_table, anno = sample_table,
  condition = "condition")
}
\arguments{
\item{input}{A dataframe containing the read counts of the genes for each sample. By default, "norm_anno" normalized read counts is used.}

\item{anno}{A sample table containing info for the samples}

\item{condition}{A character from sample table df defining the name of column in the table to use. "condition" is set by default.}
}
\value{
Returns a dataframe
}
\description{
This function provides the calculation of mean of the total reads, percent aligned etc. 
in sample table to make mean_sample_table of the average of replicates per condition.
}
