% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Motifs}
\alias{Motifs}
\alias{motifs}
\title{Motifs}
\format{A data frame}
\usage{
motifs
}
\description{
Motifs
}
\references{
This datasets corresponds to the GMT file c3.all.v6.2.entrez.gmt \url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\keyword{genes}
\keyword{pathway}
