% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{plotDEHeatmap}
\alias{plotDEHeatmap}
\title{Function to plot heatmaps of DE genes based on pheatmap}
\usage{
plotDEHeatmap(input = norm_anno, smp_table = sample_table,
  plot_anno = plot_annotation, comparison, factor,
  gene_anno = gene_annotation, conditions = "all", gene_type = "all",
  show_rownames = FALSE, cluster_cols = FALSE)
}
\description{
Function to plot heatmaps of DE genes based on pheatmap
}
