% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{plotGeneSetHeatmap}
\alias{plotGeneSetHeatmap}
\title{Heatmap of genes of specified GO, KEGG or HALLMARK gene sets}
\usage{
plotGeneSetHeatmap(cat, term, organism, show_rownames = TRUE,
  cluster_cols = FALSE)
}
\arguments{
\item{cat}{Category of the geneset. Either KEGG, GO or HALLMARK}

\item{term}{Term / ID of the geneset. Either a GO term, a KEGG pathway or a set of Hallmark genes}

\item{organism}{either "mouse" or "human"}

\item{show_rownames}{Shall the gene names be shown in the heatmap? Default is FALSE.}

\item{cluster_cols}{Shall the colums be clustered? Default is TRUE}
}
\value{
Returns a heatmap
}
\description{
This function provides a wrapper for the \link[pheatmap:pheatmap]{pheatmap}.
It can plot a subset of genes from a certain gene set based on the GO, KEGG or HALLMARK database.
}
