% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{plotGSEAHeatmap}
\alias{plotGSEAHeatmap}
\title{Function to plot a heatmap of genes responsible for gene set enrichment}
\usage{
plotGSEAHeatmap(input = norm_anno, smp_table = sample_table,
  plot_anno = plot_annotation, GSEA_result, GeneSet, term, regulation,
  show_rownames = TRUE, cluster_cols = F, gene_type = "all")
}
\description{
Function to plot a heatmap of genes responsible for gene set enrichment
}
