% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot a heatmap}
\usage{
plotHeatmap(geneset = "all", title = "", keyType = "Ensembl",
  show_rownames = FALSE, cluster_cols = FALSE)
}
\arguments{
\item{geneset}{A character vector containing the genes to plot or "all" for plotting all genes of the dataset (hardcoded to norm_anno). By default, all genes are plotted.}

\item{title}{The title of the plot}

\item{keyType}{Annotation of the genes in geneset. Either "Ensembl" or "Symbol"}

\item{show_rownames}{Shall the gene names be shown in the heatmap? Default is FALSE.}

\item{cluster_cols}{Shall the colums be clusteres? Default is FALSE}
}
\value{
Returns a heatmap
}
\description{
This function provides a wrapper for the \link[pheatmap:pheatmap]{pheatmap}.
It can plot a subset of genes, either annotated as EnsemblIDs or Gene Symbol
}
