% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plotLoadings}
\alias{plotLoadings}
\title{Function to plot heatmaps of PC loadings}
\usage{
plotLoadings(PC, ntop)
}
\arguments{
\item{PC}{principal component of which the loadings should be plotted}

\item{ntop}{number of genes to plot, sorted by variance}
}
\description{
Function to plot heatmaps of PC loadings
}
