% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_general.R
\name{plotMA}
\alias{plotMA}
\title{Custom function to plot baseMean versus fold change}
\usage{
plotMA(comparison, ylim = c(-2, 2), padjThreshold = 0.05,
  xlab = "mean of normalized counts", ylab = expression(log[2] ~ fold ~
  change), log = "x", cex = 0.45)
}
\description{
Custom function to plot baseMean versus fold change
}
