% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Function to plot a PCA}
\usage{
plotPCA(pca_input = dds_vst, ntop = 500, xPC = 1, yPC = 2, color,
  anno_colour, shape = "NULL", point_size = 3, title = "PCA",
  label = NULL, label_subset = NULL, gene_anno = gene_annotation,
  gene_type = "all")
}
\arguments{
\item{pca_input}{input dataset, default: dds_vst}

\item{ntop}{number of highest variable genes used for PCA}

\item{xPC}{Number of principal component on the x axis}

\item{yPC}{Number of principal component on the y axis}

\item{color}{Column of sample_table that defines the colour}

\item{anno_colour}{annotation of the colour values}

\item{shape}{Column of sample_table that defines the shape}

\item{point_size}{Size of points in plot}

\item{title}{Plot title}

\item{label}{Column of sample_table that defines the label}

\item{label_subset}{A character vector with a subset of labels to plot}
}
\description{
Function to plot a PCA
}
