% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_genes.R
\name{plotSingleGene}
\alias{plotSingleGene}
\title{Function to plot the normalized or batch-corrected counts for a single gene in a box plot}
\usage{
plotSingleGene(data = norm_anno, symbol, condition = "Genotype_Age",
  anno_colour = col_genotype_age, shape = NULL)
}
\description{
Function to plot the normalized or batch-corrected counts for a single gene in a box plot
}
