% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colouring.R
\name{setColours}
\alias{setColours}
\title{Creating colour annotations}
\usage{
setColours(data = plot_annotation, factorcolours = c("Set1", "Greys"),
  continouscolours = c("forestgreen"))
}
\arguments{
\item{data}{The input dataframe}

\item{factorcolours}{A vector of names of colour palettes from the RColourbrewer package. These palettes will be used to colour categorical variables}

\item{continuouscolors}{A vector of colour names that will be chosen for continuous data}
}
\value{
Returns a list with colours that can be used as an input for pheatmap
}
\description{
This functions creates colour annotations for categorical and continuous variables in a data frame to be plotted in a heatmap.
}
