% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tx_annotation_vM16_mouse}
\alias{tx_annotation_vM16_mouse}
\title{TX annotation mouse vM16}
\format{A data frame with 133944 obs. of  4 variables:
 \describe{
  \item{GENEID}{Ensembl Gene ID}
  \item{TXNAME}{Transcript ID}
  \item{SYMBOL}{Gene Symbol}
  \item{GENETYPE}{Genetype description. See \url{https://www.gencodegenes.org/pages/biotypes.html} for a list of all gene types.
  }
  }}
\usage{
tx_annotation_vM16_mouse
}
\description{
This gene annotation file is used
1) to map the Ensembl transcript IDs to Ensembl gene IDs during the tximport function and
2) annotate the Ensembl IDs with additional information such as gene symbol or type.
For consistency, this file should have been produced from the .gtf file used for building the kallisto index. It needs to consist
of four columns: Gene ENSEMBL ID, Transcript ID, Gene Symbol, Gene Type.
}
\references{
This file corresponds to ID2SYMBOL_gencode_vM16_transcript.txt, which has been generated from the GTF file gencode.vM16.primary_assembly.annotation.gtf, as downloaded from \url{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M16/gencode.vM16.primary_assembly.annotation.gtf.gz}.
}
\keyword{annotation}
