% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canonicalPathway_genes}
\alias{canonicalPathway_genes}
\title{Canonical pathway genes}
\format{A data frame
  \describe{
  \item{ont}{Pathway term}
  \item{gene}{Entrez ID}
  }}
\usage{
canonicalPathway_genes
}
\description{
List with KEGG canonical pathways from the MiSigDB gene set.
}
\references{
This datasets corresponds to c2.cp.v6.2.entrez.gmt \url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\keyword{datasets}
