% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{plotDEHeatmap}
\alias{plotDEHeatmap}
\title{Function to plot heatmaps of DE genes based on pheatmap}
\usage{
plotDEHeatmap(comparison, factor, conditions = "all",
  show_rownames = FALSE, cluster_cols = FALSE)
}
\description{
Function to plot heatmaps of DE genes based on pheatmap
}
