% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{plotGSEAHeatmap}
\alias{plotGSEAHeatmap}
\title{Function to plot a heatmap of genes responsible for gene set enrichment}
\usage{
plotGSEAHeatmap(GSEA_result, GeneSet, term, regulation)
}
\description{
Function to plot a heatmap of genes responsible for gene set enrichment
}
